+++ Routes/campgrounds.js
@@ -1,0 +1,74 @@
+const express = require('express');
+const router = express.Router();
+const expressError = require('../utils/expressError');
+const Campground = require('../models/campground');
+const catchAsync = require('../utils/catchAsync');
+const { campgroundSchema, reviewSchema } = require('../schemas')
+
+//middleware
+const validateCampground = (req, res, next) => {
+    const { error } = campgroundSchema.validate(req.body);
+    if (error) {
+        const message = error.details.map(el => el.message).join(',');
+        throw new expressError(message, 400);
+    }
+    else {
+        next();
+    }
+}
+
+
+
+//index route
+router.get('/', catchAsync(async (req, res) => {
+
+    const campgrounds = await Campground.find({});
+    res.render('campgrounds/index', { campgrounds });
+}))
+
+//create new route begins
+router.get('/new', (req, res) => {
+    res.render('campgrounds/new');
+})
+
+router.post('/', validateCampground, catchAsync(async (req, res) => {
+
+    const campground = new Campground(req.body.campground);
+    await campground.save();
+    res.redirect(`/campgrounds/${campground._id}`)
+}))
+
+// end of new route
+
+//show route 
+router.get('/:id', catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id).populate('reviews');
+    res.render('campgrounds/show', { campground });
+}))
+
+//route to edit begins
+
+router.get('/:id/edit', catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    res.render('campgrounds/edit', { campground });
+}))
+
+router.put('/:id', validateCampground, catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    res.redirect(`/campgrounds/${campground.id}`)
+}))
+
+//route to edit ends
+
+//route to delete begins
+
+router.delete('/:id', catchAsync(async (req, res) => {
+    const { id } = req.params;
+    await Campground.findByIdAndDelete(id);
+    res.redirect('/campgrounds');
+}))
+
+//route to delete ends
+
+module.exports = router;
