+++ Routes/reviews.js
@@ -1,0 +1,41 @@
+const express = require('express');
+const router = express.Router({ mergeParams: true });
+const expressError = require('../utils/expressError');
+const catchAsync = require('../utils/catchAsync');
+
+//models
+const Campground = require('../models/campground');
+const Review = require('../models/review')
+
+const { reviewSchema } = require('../schemas')
+
+//middleware to validate data
+const validateReview = (req, res, next) => {
+    const { error } = reviewSchema.validate(req.body);
+    if (error) {
+        const message = error.details.map(el => el.message).join(',');
+        throw new expressError(message, 400);
+    }
+    else {
+        next();
+    }
+}
+
+//reviews routes
+router.post('/', validateReview, catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    const review = new Review(req.body.review);
+    campground.reviews.push(review);
+    await review.save();
+    await campground.save();
+    res.redirect(`/campgrounds/${campground._id}`)
+}));
+
+router.delete('/:reviewId', catchAsync(async (req, res) => {
+    const { id, reviewId } = req.params;
+    await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
+    await Review.findByIdAndDelete(req.params.reviewId);
+    res.redirect(`/campgrounds/${id}`);
+}))
+
+module.exports = router;
