+++ app.js
@@ -2,14 +2,15 @@
 const express = require('express');
 const path = require('path');
 const mongoose = require('mongoose');
-const Campground = require('./models/campground')
 const methodOverride = require('method-override');
 const ejsMate = require('ejs-mate');
-const catchAsync = require('./utils/catchAsync');
 const expressError = require('./utils/expressError')
-const { campgroundSchema, reviewSchema } = require('./schemas')
-const Review = require('./models/review');
 
+//require the routes
+const campgrounds = require('./Routes/campgrounds');
+const reviews = require('./Routes/reviews');
+
+//database connection settings
 mongoose.connect('mongodb://localhost:27017/yelp-camp', {
     useNewUrlParser: true,
     useUnifiedTopology: true,
@@ -23,115 +24,24 @@ db.once('open', () => {
     console.log("we're connected!");
 });
 
-//middleware
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    if (error) {
-        const message = error.details.map(el => el.message).join(',');
-        throw new expressError(message, 400);
-    }
-    else {
-        next();
-    }
-}
 
-const validateReview = (req, res, next) => {
-    const { error } = reviewSchema.validate(req.body);
-    if (error) {
-        const message = error.details.map(el => el.message).join(',');
-        throw new expressError(message, 400);
-    }
-    else {
-        next();
-    }
-}
-
-
-//set and UISE
+//set and USE
+app.engine('ejs', ejsMate)
 app.set('view engine', 'ejs');
 app.set('views', path.join(__dirname, 'views'));
 app.use(express.urlencoded({ extended: true }))
 app.use(methodOverride('_method'));
-app.engine('ejs', ejsMate)
+
 mongoose.set('useFindAndModify', false);
 
-
+//use the routes
+app.use("/campgrounds", campgrounds)
+app.use("/campgrounds/:id/reviews", reviews)
 
 //routes
 app.get('/', (req, res) => {
-    res.redirect('/campgrounds')
+    res.redirect('home')
 })
-
-//index route
-app.get('/campgrounds', catchAsync(async (req, res) => {
-
-    const campgrounds = await Campground.find({});
-    res.render('campgrounds/index', { campgrounds });
-}))
-
-//create new route begins
-app.get('/campgrounds/new', (req, res) => {
-    res.render('campgrounds/new');
-})
-
-app.post('/campgrounds', validateCampground, catchAsync(async (req, res) => {
-
-    const campground = new Campground(req.body.campground);
-    await campground.save();
-    res.redirect(`/campgrounds/${campground._id}`)
-}))
-
-// end of new route
-
-//show route 
-app.get('/campgrounds/:id', catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews');
-    res.render('campgrounds/show', { campground });
-}))
-
-//route to edit begins
-
-app.get('/campgrounds/:id/edit', catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    res.render('campgrounds/edit', { campground });
-}))
-
-app.put('/campgrounds/:id', validateCampground, catchAsync(async (req, res) => {
-    const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
-    res.redirect(`/campgrounds/${campground.id}`)
-}))
-
-//route to edit ends
-
-//reviews
-app.post('/campgrounds/:id/reviews', validateReview, catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    const review = new Review(req.body.review);
-    campground.reviews.push(review);
-    await review.save();
-    await campground.save();
-    res.redirect(`/campgrounds/${campground._id}`)
-}))
-
-app.delete('/campgrounds/:id/reviews/:reviewId', catchAsync(async (req, res) => {
-    const { id, reviewId } = req.params;
-    await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-    await Review.findByIdAndDelete(req.params.reviewId);
-    res.redirect(`/campgrounds/${id}`);
-}))
-
-
-//route to delete begins
-
-app.delete('/campgrounds/:id', catchAsync(async (req, res) => {
-    const { id } = req.params;
-    await Campground.findByIdAndDelete(id);
-    res.redirect('/campgrounds');
-}))
-
-//route to delete ends
-
 
 //error handler
 app.all('*', (req, res, next) => {
